pyproject.toml
[project]
name = "clean-bot"
version = "0.1.0"
description = "LINE Bot for managing periodic cleaning tasks"
requires-python = ">=3.11.7"
dependencies = [
    "boto3>=1.42.38",
    "line-bot-sdk>=3.22.0",
]

[dependency-groups]
dev = [
    "moto>=5.1.20",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
